% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gsvcm.R
\name{cv.gsvcm}
\alias{cv.gsvcm}
\title{k-fold cross-validation MSPE for generalized spatially varying coefficient regression}
\usage{
cv.gsvcm(y, X, S, V, Tr, d = 2, r = 1, lambda = 10^seq(-6, 6, by =
  0.5), family, off = 0, r.theta = c(2, 8), nfold = 10,
  initial = 123, eps.sigma = 0.01, method = "GSVCM", Cp = TRUE)
}
\arguments{
\item{y}{The response of dimension \code{n} by one, where \code{n} is the number of observations.
\cr}

\item{X}{The design matrix of dimension \code{n} by \code{p}, with an intercept. Each row is an observation vector.
\cr}

\item{S}{The cooridinates of dimension \code{n} by two. Each row is the coordinates of an observation.
\cr}

\item{V}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{d}{The degree of piecewise polynomials -- default is 2.
\cr}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.
\cr}

\item{lambda}{The vector of the candidates of penalty parameter -- default is grid points of 10 to the power of a sequence from -6 to 6 by 0.5.
\cr}

\item{family}{The family object, specifying the distribution and link to use.
\cr}

\item{off}{offset -- default is 0.
\cr}

\item{r.theta}{The endpoints of an interval to search for an additional parameter \code{theta} for negative binomial scenario -- default is c(2,8).
\cr}

\item{nfold}{The number of folds -- default is 10. Although \code{nfold} can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable for \code{nfolds} is 3.
\cr}

\item{initial}{The seed used for cross-validation sample -- default is 123.
\cr}

\item{eps.sigma}{Error tolerance for the Pearson estimate of the scale parameter, which is as close as possible to 1, when estimating an additional parameter \code{theta} for negative binomial scenario -- default is 0.01.
\cr}

\item{method}{GSVCM or GSVCMQR. GSVCM is based on Algorithm 1 in Subsection 3.1 and GSVCMQR is based on Algorithm 2 in Subsection 3.2 -- default is GSVCM.
\cr}

\item{Cp}{TRUE or FALSE. There are two modified measures based on the QRGSVCM method for smoothness parameters in the manuscript. TRUE is for Cp measure and FALSE is for GCV measure.
\cr}
}
\value{
The k-fold cross-validation (CV) mean squared prediction error (MSPE).
\cr
}
\description{
\code{cv.gsvcm} implements k-fold cross-validation MSPE for generalized spartially varying coefficient regression, and returns the mean squared prediction error (MSPE).
}
\details{
This R package is the implementation program for manuscript entitled "Generalized Spatially Varying Coefficinet Models" by Myungjin Kim and Li Wang.
}
\examples{
# See an example of fit.gsvcm.
}
