% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datagenerator.R
\name{Datagenerator}
\alias{Datagenerator}
\title{Generating samples in Simulation.}
\usage{
Datagenerator(family, grid = 0.005)
}
\arguments{
\item{family}{The family object, specifying the distribution and link to use. Choose "gaussian" for Gaussian distribution, "poisson" for poisson distribution, and "nb_bps" for negative binomial distribution.
\cr}

\item{grid}{The distance between grid points -- default is set to 0.005.
\cr}
}
\value{
A data matrix with a response ('y'), true coefficient functions ('m1' and 'm2'), covariates ('x1' and 'x2'), and locations ('u' and 'v').
\cr
}
\description{
\code{Datagenerator} is used to generate samples on horseshoe domain for Scenario 1 (Gaussian), Scenario 2 (Poisson), and Scenario 3 (negative binomial).
}
\details{
This R package is the implementation program for manuscript entitled "Generalized Spatially Varying Coefficinet Models" by Myungjin Kim and Li Wang.
}
