% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gsvcm.R
\name{gsvcm.est.reduce}
\alias{gsvcm.est.reduce}
\title{Estimation for reduced model (GSVCMs with partially linear part)}
\usage{
gsvcm.est.reduce(y, X.l, X.nl, BQ2, P, lambda, family, off = 0,
  theta = 0)
}
\description{
This is an internal function of package \code{gsvcm}.
}
