% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gsvcm.R
\name{gsvcm.GQLR}
\alias{gsvcm.GQLR}
\title{Quasi-likelihood test statistics based on a variety of \code{family}, evaluated by full and reduced estimates.}
\usage{
gsvcm.GQLR(yi, y_hat1, y_hat2, family, theta)
}
\description{
This is an internal function of package \code{gsvcm}.
}
