% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvcm.R
\docType{package}
\name{gsvcm}
\alias{gsvcm}
\alias{gsvcm-package}
\title{gsvcm:	A	package	for	generalized spatially varying coefficient models.}
\description{
This R package is the implementation program for manuscript entitled "Generalized Spatially Varying Coefficinet Models" by Myungjin Kim and Li Wang.
The	gsvcm	package	provides the following functions:
fit.gsvcm, test.gsvcm, and cv.gsvcm.
}
\section{fit.gsvcm}{

The \code{\link{fit.gsvcm}}  fits the generalized spatially varying coeffcient models.
 \cr
}

\section{test.gsvcm}{

The \code{\link{test.gsvcm}}  implements Generalized Quasi-Likelihood Ratio (GQLR) test using wild bootstrap.
 \cr
}

\section{cv.gsvcm}{

The \code{\link{cv.gsvcm}}  implements k-fold cross-validation MSPE for generalized spartially varying coefficient regression, and returns the mean squared prediction error (MSPE).
 \cr
}

