% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvcm_est.R
\name{gsvcm_est}
\alias{gsvcm_est}
\title{Estimation for GSVCMs and QRGSVCMs, which is used in R funtion "fit.gsvcm".}
\usage{
gsvcm_est(y, X, BQ2, P, lambda, family, off = 0, theta = 0)
}
\description{
Estimation for GSVCMs and QRGSVCMs, which is used in R funtion "fit.gsvcm".
}
