% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gsvcm.R
\name{plot.gsvcm}
\alias{plot.gsvcm}
\title{Produces coefficient function plots for a fitted "gsvcm" object.}
\usage{
\method{plot}{gsvcm}(mfit, gridnumber = 50)
}
\arguments{
\item{mfit}{Fitted ``gsvcm" object.
\cr}

\item{gridnumber}{the number of grid points for plots.
\cr}
}
\value{
None
}
\description{
This function produces the plots of the estimated coefficient function for an generalized spatially varying coefficient regression based on sample locations.
}
