% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gsvcm.R
\name{predict.gsvcm}
\alias{predict.gsvcm}
\title{Making predictions from a fitted generalized spatially varying coefficient model}
\usage{
predict.gsvcm(mfit, Xpred, Spred)
}
\arguments{
\item{mfit}{Fitted ``gsvcm" object.
\cr}

\item{Xpred}{The design matrix for prediction.
\cr}

\item{Spred}{The cooridinates for prediction.
\cr}
}
\value{
A vector of predicted response is returned.
\cr
}
\description{
This function is used to make predictions of the generalized spatially varying coefficient models.
}
\details{
This R package is the implementation program for manuscript entitled "Generalized Spatially Varying Coefficinet Models" by Myungjin Kim and Li Wang.
\cr
}
\examples{
# See an example of fit.gsvcm.
}
