% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gsvcm.R
\name{test.gsvcm}
\alias{test.gsvcm}
\title{Generalized Quasi-Likelihood Ratio (GQLR) test for generalized spatially varying coefficient regression}
\usage{
test.gsvcm(y, X, S, V, Tr, d, r, lambda = 10^seq(-6, 6, by = 0.5),
  test_iter = 1, family, off = 0, r.theta = c(2, 8), nB = 100,
  initial = 123, eps.sigma = 0.01)
}
\arguments{
\item{y}{The response of dimension \code{n} by one, where \code{n} is the number of observations.
\cr}

\item{X}{The design matrix of dimension \code{n} by \code{p}, with an intercept. Each row is an observation vector.
\cr}

\item{S}{The cooridinates of dimension \code{n} by two. Each row is the coordinates of an observation.
\cr}

\item{V}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{d}{The degree of piecewise polynomials -- default is 2.
\cr}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.
\cr}

\item{lambda}{The vector of the candidates of penalty parameter -- default is grid points of 10 to the power of a sequence from -6 to 6 by 0.5.
\cr}

\item{test_iter}{The function to be tested -- default is the first function, which corresponds to the first covariate.
\cr}

\item{family}{The family object, specifying the distribution and link to use.
\cr}

\item{off}{offset -- default is 0.
\cr}

\item{r.theta}{smoothness penalty parameter candidates -- default is c(2,8).
\cr}

\item{nB}{The number of boostrap replication -- default is 100.
\cr}

\item{initial}{The seed used for bootstrap sample in the GQLR test -- default is 123.
\cr}

\item{eps.sigma}{Error tolerance for the Pearson estimate of the scale parameter, which is as close as possible to 1, when estimating an additional parameter \code{theta} for negative binomial scenario -- default is 0.01.
\cr}
}
\value{
The function returns a list with the following items:
\item{b.GQLR}{The GQLR test statistics based on sample.}
\item{obs.GQLR}{The GQLR test statistics based on each boostrap iteration.}
\item{pvalue}{The p-value for the GQLR test.}
}
\description{
\code{test.gsvcm} implements Generalized Quasi-Likelihood Ratio (GQLR) test using wild bootstrap.
}
\details{
This R package is the implementation program for manuscript entitled "Generalized Spatially Varying Coefficinet Models" by Myungjin Kim and Li Wang.
\cr
}
\examples{
# See an example of fit.gsvcm.

}
